requirements.txt
python-dotenv~=1.0.1
nltk~=3.8.1
numpy~=1.26.4
pandas~=2.2.2
scikit-learn~=1.4.2
kaggle~=1.6.12
tensorflow~=2.15.0
requests~=2.31.0
langchain~=0.1.20
langchain-core~=0.1.52
langchain-openai~=0.1.7
langchain-community~=0.0.38
chromadb~=0.5.0
fastapi~=0.111.0
pydantic~=2.7.1
gradio~=4.31.4
